feat: TASK 6 — redesign win condition (gameplay-first scoring)
Old formula: (level * 10) + TT — optimizer-friendly, no leadership signal
New formula: (level * 10) + (TT * 1.5, cap 15) + reputation(±5) + diversity(0-5) - selfish_tax(0-10)

Key changes:
- Reputation matters: selfish play now has a real cost
- Leadership diversity bonus: rewards versatile leaders over stat-hoarders
- Selfish tax: breaking more promises than keeping = up to -10 points
- TT weighted at 1.5x instead of 1:1 to prevent pure TT optimization

New badges:
- The Carrier: Summit + TT>=8 + reputation>=0 + net positive promises
- The Catalyst: Did NOT summit + highest TT + positive cross-player effects
- The Strategist: 4+ distinct leadership behaviors demonstrated
- Solo Peak: Summit + TT<8 or negative reputation
- Climber: Default

Badge priority ranking: Carrier > Catalyst > Strategist > SoloPeak > Climber
Updated: Badge enum, GamePlayer (calculateScore + badge qualifiers),
GameService (finishGame + assignBadge), config, views

+++ app/Models/GameResult.php
@@ -30,7 +30,7 @@ class GameResult extends Model
             'final_mp'          => 'integer',
             'final_sp'          => 'integer',
             'final_tt'          => 'integer',
-            'final_score'       => 'integer',
+            'final_score'       => 'float',
             'rank'              => 'integer',
             'final_reputation'  => 'integer',
             'final_resources'  => 'integer',
